+++ src/common.rs
@@ -4,23 +4,24 @@ use crate::expand::{
     BRACE_BEGIN, BRACE_END, BRACE_SEP, BRACE_SPACE, HOME_DIRECTORY, INTERNAL_SEPARATOR,
     PROCESS_EXPAND_SELF, PROCESS_EXPAND_SELF_STR, VARIABLE_EXPAND, VARIABLE_EXPAND_SINGLE,
 };
-use crate::global_safety::AtomicRef;
-use crate::global_safety::RelaxedAtomicBool;
-use crate::prelude::*;
-use crate::terminal::Outputter;
-use crate::termsize::Termsize;
-use crate::wildcard::{ANY_CHAR, ANY_STRING, ANY_STRING_RECURSIVE};
-use crate::wutil::fish_iswalnum;
+use crate::{
+    global_safety::{AtomicRef, RelaxedAtomicBool},
+    prelude::*,
+    terminal::Outputter,
+    termsize::Termsize,
+    wildcard::{ANY_CHAR, ANY_STRING, ANY_STRING_RECURSIVE},
+    wutil::fish_iswalnum,
+};
 use fish_fallback::fish_wcwidth;
 use fish_feature_flags::{FeatureFlag, feature_test};
-use fish_wcstringutil::wcs2bytes;
-use fish_widestring::{decode_byte_from_char, encode_byte_to_char, subslice_position};
+use fish_widestring::{decode_byte_from_char, encode_byte_to_char, subslice_position, wcs2bytes};
 use nix::sys::termios::Termios;
-use std::env;
-use std::ffi::{CStr, OsStr};
-use std::os::unix::prelude::*;
-use std::sync::atomic::Ordering;
-use std::sync::{MutexGuard, OnceLock};
+use std::{
+    env,
+    ffi::{CStr, OsStr},
+    os::unix::prelude::*,
+    sync::{MutexGuard, OnceLock, atomic::Ordering},
+};
 
 use fish_common::*;
