+++ src/screen.rs
@@ -26,8 +26,8 @@ use crate::termsize::Termsize;
 use crate::wutil::fstat;
 use fish_common::write_loop;
 use fish_fallback::{fish_wcswidth_canonicalizing, fish_wcwidth};
-use fish_wcstringutil::{fish_wcwidth_visible, string_prefixes_string, wcs2bytes};
-use fish_widestring::ELLIPSIS_CHAR;
+use fish_wcstringutil::{fish_wcwidth_visible, string_prefixes_string};
+use fish_widestring::{ELLIPSIS_CHAR, wcs2bytes};
 use libc::{STDERR_FILENO, STDOUT_FILENO};
 use nix::sys::termios;
 use std::cell::RefCell;
