+++ src/universal_notifier/kqueue.rs
@@ -3,7 +3,7 @@ use crate::flogf;
 use crate::prelude::*;
 use crate::universal_notifier::UniversalNotifier;
 use crate::wutil::wdirname;
-use fish_wcstringutil::wcs2osstring;
+use fish_widestring::wcs2osstring;
 use nix::sys::event::{EvFlags, EventFilter, FilterFlag, KEvent, Kqueue};
 use std::fs::File;
 use std::os::fd::AsFd;
